removed unused documentPositionCalculator and move astNodeUtils under utils
