+++ src/languageservice/parser/schemaValidation/baseValidator.ts
@@ -25,7 +25,7 @@ import * as l10n from '@vscode/l10n';
 import { URI } from 'vscode-uri';
 import { Diagnostic, DiagnosticSeverity, Range } from 'vscode-languageserver-types';
 import type { TextDocument } from 'vscode-languageserver-textdocument';
-import { contains, getNodeValue } from '../astNodeUtils';
+import { contains, getNodeValue } from '../../utils/astNodeUtils';
 import { getValidator } from './validatorFactory';
 
 export const YAML_SOURCE = 'YAML';
