+++ src/languageservice/utils/astNodeUtils.ts
